refactor(ui): give the game-detail panel's event lane its own module
The panel's `[appId]` effect held eleven listeners, their dispatcher and the listener wiring as
closures nested two levels inside the component. That nesting is most of what the file's suppressed
cognitive-complexity score is made of: a guard costs 1 at module scope and 3 in there.

They now live at module scope in panelEvents.ts and take an explicit context instead of sharing the
effect's closure. The effect keeps the loadData call and one wiring call whose teardown unsubscribes
all four listeners. PanelState, RomBinding and bindRom move to panelState.ts alongside the reads
that fold answers into them, so the panel and the event lane can both reach them without an import
cycle. The panel drops from 1078 lines to 527.

Two properties of the closure had to survive the move. `cancelled` is carried as a function rather
than a boolean, because it is read when an answer lands rather than when the read was issued; and
`bindRom` still compares `romIdRef` inside the write, so a bound fold is still refused at land time.
The dispatcher returns the chosen handler's promise instead of awaiting it, so the caller's await
resolves on the turn it did before.

Nothing else moved. The ten bound write sites are the same ten functions, the six raw-setter writes
are the same six, handleCoreChange still folds its two reads into one write, and handleBiosChange is
unchanged including the unbound write #1718 tracks.

The rom-binding invariant register pointed at the panel file for writes that are no longer in it. It
now names the two modules that hold them.

Deliberately not here: the render blocks and the lazy SAVES effect stay in the component, and so
does the S3776 suppression. It can only come off once the body holds nothing but the render, which
is the last step in this chain.

RomMGameInfoPanel.test.tsx is unchanged, and the suite reports the same file and test counts as
before the move.

docs: N/A — pure refactor; nothing the user sees changes.

Part of #994.
